requirements.txt
mistralai
python-dotenv
langsmith
langchain_core